Fix Vue.js functionality due to malformed public-url in Windows
The current project builds flawlessly in Mac/Linux, but loses Vue.js functionality while executing the build in Windows. The apostrophes present in the public-url in package.json are encoded incorrectly as "%27" while building in Windows and thus invalidate the resulting url. These apostrophes are unnecessary and can be safely removed without affecting the build process in any OS.